Desktop/FORGE-main/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "forgeagent"
version = "3.0.0"
description = "Local AI coding agent hub — train, deploy, and manage offline Ollama models"
requires-python = ">=3.10"
dependencies = [
    "rich>=13.7",
    "textual>=0.50",
    "httpx>=0.27",
    "click>=8.1",
    "python-dotenv>=1.0",
    "pydantic>=2.6",
]

[project.scripts]
forgeagent = "forgeagent.__main__:main"

[tool.setuptools.packages.find]
where = ["."]
